back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Algerdi <kzlv.vldslv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
channels = "^3.0.4"
channels-redis = "^3.4.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
